+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Breaking:
 - Removed Command `PgPepClearAnnotationsCommand`
 - Removed Command `PgPepAnnotateCommand`
 - Removed Commands `PgPepClearHighlightedCommand` and `PgPepToggleHighlightCommand`
+- Removed Command `PgPepViewNamespaceStatusCommand` and its settings:
+	- `view_status_show_namespace`
+	- `view_status_show_namespace_prefix`
+	- `view_status_show_namespace_suffix`
 
 ## 0.19.0 - 2023-10-31
 There were many internal changes in this version and I'm sorry I made some breaking changes.